temperature_service/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0

# Async support
asyncio==3.4.3
aiohttp==3.8.5
httpx==0.25.0

# Database
influxdb==5.3.1
aioinflux==0.9.0
redis==5.0.1
aioredis==2.0.1

# Observability
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-redis==0.41b0
prometheus-client==0.17.1
structlog==23.1.0

# Utilities
tenacity==8.2.3
ujson==5.8.0
msgpack==1.0.5
orjson==3.9.7

# Development and testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0
black==23.9.1
isort==5.12.0
mypy==1.5.1
flake8==6.1.0
